integrations/pyproject.toml
[project]
name = "llm_integrations"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",

  # For simple contract files in golden tests (nice for readability)
  "pyyaml>=6.0",

  # Optional, but very handy if you do DB assertions directly from integrations
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",

  # Optional, but makes env/config handling painless
  "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
# Match the proposed folder structure
testpaths = ["test_server_live", "test_eval_job", "test_policy_job", "test_e2e_loop"]

# Make markers explicit (pytest will warn/error if you use undeclared markers)
markers = [
  "server_live: live server contract tests",
  "eval_job: eval job integration tests",
  "policy_job: policy job integration tests",
  "e2e_loop: end-to-end server/eval/policy workflow tests",
  "generate_only: generate-only deployment contract tests",
  "full: extract-enabled/full-capability contract tests",
  "requires_db: tests that assert DB persistence (InferenceLog, cache tables, etc.)",
  "requires_redis: tests that assert redis cache hit/miss behavior",
  "requires_metrics: tests that assert /metrics increments",
  "slow: slower integration tests (concurrency, larger payloads, etc.)",
]

# Sensible defaults; override locally as needed:
# -q keeps output compact
# -ra shows summary of skipped/failed
# --strict-markers prevents typos from silently creating new markers
addopts = "-q -ra --strict-markers"

# asyncio defaults (since you have concurrency goals)
asyncio_mode = "auto"

[tool.uv]
package = false
